Changes related with Error handling and some unusual cases (see desc.)
Changelog:
Enhanced a logic related with "start_block_offset" and case in which there's no gathered block in "data"

Added a Logic which will pass to the second query If 1st returned 0 results.

Deleted "gather_to_start.py" as It's now not needed at all in all cases.
Deleted "fetch_latest_block.sql" as It's now not needed.